refactor(extension/podman): extract DarwinSocketCompatibility to dedicated file
